improvements to books.js and renaming of appcontext
