+++ pruebas/01-reading-list/blasferna/src/app/page.js
@@ -1,4 +1,3 @@
-import DATA from "@/books.json";
 import { AvailableList } from "@/components/book-list";
 import Navigation from "@/components/navigation";
 
@@ -7,7 +6,7 @@ export default function Home() {
   return (
     <>
       <Navigation></Navigation>
-      <AvailableList data={DATA}></AvailableList>
+      <AvailableList></AvailableList>
     </>
   );
 }
