+++ pruebas/01-reading-list/blasferna/src/lib/books.js
@@ -1,3 +1,5 @@
+import DATA from "@/books.json";
+
 const READING_LIST_STORAGE_KEY = "readingList";
 
 const groupByGenre = (data) => {
@@ -13,14 +15,16 @@ const groupByGenre = (data) => {
 
 const getInReadingList = () => {
   if (typeof window !== "undefined") {
-    return JSON.parse(localStorage.getItem(READING_LIST_STORAGE_KEY) || "[]") || [];
+    return (
+      JSON.parse(localStorage.getItem(READING_LIST_STORAGE_KEY) || "[]") || []
+    );
   }
   return [];
 };
 
-const getReadingList = (data) => {
+const getReadingList = () => {
   const readingList = getInReadingList();
-  const filteredList = data.library.filter((element) =>
+  const filteredList = DATA.library.filter((element) =>
     readingList.includes(element.book.ISBN)
   );
   return groupByGenre(filteredList);
@@ -30,16 +34,16 @@ const getReadingListCount = () => {
   return getInReadingList().length;
 };
 
-const getAvailableList = (data) => {
+const getAvailableList = () => {
   const readingList = getInReadingList();
-  const filteredList = data.library.filter(
+  const filteredList = DATA.library.filter(
     (element) => !readingList.includes(element.book.ISBN)
   );
   return groupByGenre(filteredList);
 };
 
-const getAvailableListCount = (data) => {
-  return getAvailableList(data).length;
+const getAvailableListCount = () => {
+  return DATA.library.length - getInReadingList().length;
 };
 
 const isInReadingList = (isbn) => {
@@ -59,6 +63,26 @@ const removeFromReadingList = (isbn) => {
   localStorage.setItem(READING_LIST_STORAGE_KEY, JSON.stringify(newList));
 };
 
+const getAuthorOtherBooks = (book) => {
+  const otherBooks = [];
+  const array = book.author.otherBooks;
+  for (let index = 0; index < array.length; index++) {
+    const element = array[index];
+    try {
+      otherBooks.push(getByTitle(element));
+    } catch (error) {}
+  }
+  return otherBooks;
+};
+
+const getByISBN = (isbn) => {
+  return DATA.library.find((obj) => obj.book.ISBN === isbn).book;
+};
+
+const getByTitle = (title) => {
+  return DATA.library.find((obj) => obj.book.title === title).book;
+};
+
 export {
   READING_LIST_STORAGE_KEY,
   addToReadingList,
@@ -70,4 +94,7 @@ export {
   groupByGenre,
   isInReadingList,
   removeFromReadingList,
+  getAuthorOtherBooks,
+  getByISBN,
+  getByTitle,
 };
